+++ .npmignore
@@ -5,7 +5,7 @@
 !packages/*/src/**/*
 !packages/*/index.js
 !prebuilds/**/*
-!proto/**/*
+!protobuf/**/*
 !scripts/**/*
 !CONTRIBUTING.md
 !LICENSE
