+++ packages/dd-trace/src/profiling/encoders/pprof.js
@@ -3,7 +3,7 @@
 const util = require('util')
 const zlib = require('zlib')
 const gzip = util.promisify(zlib.gzip)
-const { perftools } = require('../../../../../proto/profile')
+const { perftools } = require('../../../../../protobuf/profile')
 const { Profile } = perftools.profiles
 
 class Encoder {
